Set `package-mode = false` for Poetry project
Attempting to build the `Dockerfile` using the latest version of Poetry results in an error that suggests either passing the flags `--no-root --no-directory` to Poetry or setting the project's `package-mode` to `false` in the project configuration file.

I believe that the best solution is to change the package mode, as this project is an application and should be indicated as such to Poetry. However, I did come across a page in the Poetry docs which encourages using the above-mentioned flags when deploying with Docker anyway. This reduces the chance of cache busting. This may be worth looking into at some point.

[Poetry Docs](https://python-poetry.org/docs/faq/#poetry-busts-my-docker-cache-because-it-requires-me-to-copy-my-source-files-in-before-installing-3rd-party-dependencies)